+++ examples/02_typescript/src/DisplayData.tsx
@@ -1,5 +1,6 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-// @ts-ignore
+// eslint-disable-next-line spaced-comment
+/// <reference types="react/experimental" />
+
 import React, { useTransition } from 'react';
 
 type Props = {