+++ examples/02_typescript/src/Item.tsx
@@ -1,15 +1,13 @@
 import React, { useState } from 'react';
 
-import { prefetch, useFetch } from 'react-hooks-fetch';
+import { useFetch } from 'react-hooks-fetch';
 
-import { fetchFunc } from './fetchFuncs';
+import { store } from './fetchStore';
 import DisplayData from './DisplayData';
-
-const suspendable = prefetch(fetchFunc, '1');
 
 const Item: React.FC = () => {
   const [id, setId] = useState('1');
-  const { result, refetch } = useFetch(suspendable);
+  const [result, refetch] = useFetch(store, '1' as string);
   return (
     <div>
       User ID: <input value={id} onChange={(e) => setId(e.target.value)} />
@@ -19,7 +17,7 @@ const Item: React.FC = () => {
 };
 
 export const ErrorItem: React.FC = () => {
-  const { result, refetch } = useFetch(suspendable);
+  const [result, refetch] = useFetch(store, '1' as string);
   return (
     <div>
       <DisplayData id="-1" result={result} refetch={refetch} />