+++ examples/02_typescript/src/index.ts
@@ -1,10 +1,11 @@
+// eslint-disable-next-line spaced-comment
+/// <reference types="react-dom/experimental" />
+
 import React from 'react';
-import {
-  // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-  // @ts-ignore
-  createRoot,
-} from 'react-dom';
+import { createRoot } from 'react-dom';
 
 import App from './App';
 
-createRoot(document.getElementById('app')).render(React.createElement(App));
+const ele = document.getElementById('app');
+if (!ele) throw new Error('no app');
+createRoot(ele).render(React.createElement(App));