+++ src/index.ts
@@ -1,3 +1,3 @@
 export { ErrorBoundary } from './ErrorBoundary';
-export { useFetch, useFetchWithoutPrefetch } from './useFetch';
-export { prefetch } from 'react-suspense-fetch';
+export { createFetchStore } from './createFetchStore';
+export { useFetch } from './useFetch';